+++ src/Components/Routes.tsx
@@ -9,6 +9,8 @@ import BucketDetail from "../Views/BucketPanel/BucketDetail";
 import TokenList from "../Views/SecurityPanel/TokenList";
 import TokenDetail from "../Views/SecurityPanel/TokenDetail";
 import {TokenPermissions} from "reduct-js";
+import Replications from "../Views/Replications/Replications";
+import ReplicationDetail from "../Views/Replications/ReplicationDetail";
 
 interface Props extends RouteComponentProps {
     backendApi: IBackendAPI;
@@ -21,8 +23,8 @@ interface Props extends RouteComponentProps {
 function PrivateRoute({children, authorized, ...rest}) {
     return (
         <Route {...rest} render={() => {
-            return authorized ? children : <Redirect to="/login"/>;
-        }}/>
+            return authorized ? children : <Redirect to="/login" />;
+        }} />
     );
 }
 
@@ -32,34 +34,42 @@ export function Routes(props: Props): JSX.Element {
     return (
         <Switch>
             <Route exact path="/">
-                {authorized ? <Redirect to="/dashboard"/> : <Redirect to="/login"/>}
+                {authorized ? <Redirect to="/dashboard" /> : <Redirect to="/login" />}
             </Route>
 
             <Route exact path="/login">
-                <Login {...props}/>
+                <Login {...props} />
             </Route>
 
             <PrivateRoute exact path="/dashboard" authorized={authorized}>
-                <Dashboard {...props}/>
+                <Dashboard {...props} />
             </PrivateRoute>
 
             <PrivateRoute exact path="/buckets" authorized={authorized}>
-                <BucketList client={props.backendApi.client} {...props}/>
+                <BucketList client={props.backendApi.client} {...props} />
             </PrivateRoute>
 
             <PrivateRoute exact path="/buckets/:name" authorized={authorized}>
-                <BucketDetail client={props.backendApi.client} {...props}/>
+                <BucketDetail client={props.backendApi.client} {...props} />
+            </PrivateRoute>
+
+            <PrivateRoute exact path="/replications" authorized={authorized}>
+                <Replications client={props.backendApi.client} {...props} />
+            </PrivateRoute>
+
+            <PrivateRoute exact path="/replications/:name" authorized={authorized}>
+                <ReplicationDetail client={props.backendApi.client} {...props} />
             </PrivateRoute>
 
             <PrivateRoute exact path="/tokens" authorized={authorized && props.permissions?.fullAccess}>
-                <TokenList client={props.backendApi.client}/>
+                <TokenList client={props.backendApi.client} />
             </PrivateRoute>
 
             <PrivateRoute exact path="/tokens/:name" authorized={authorized && props.permissions?.fullAccess}>
-                <TokenDetail client={props.backendApi.client}/>
+                <TokenDetail client={props.backendApi.client} />
             </PrivateRoute>
             <Route>
-                <NotFount/>
+                <NotFount />
             </Route>
         </Switch>
     );